+++ openeo/rest/datacube.py
@@ -2361,6 +2361,10 @@ class DataCube(_ProcessGraphAbstraction):
         outputfile: Optional[Union[str, pathlib.Path]] = None,
         out_format: Optional[str] = None,
         *,
+        title: Optional[str] = None,
+        description: Optional[str] = None,
+        plan: Optional[str] = None,
+        budget: Optional[float] = None,
         print: typing.Callable[[str], None] = print,
         max_poll_interval: float = 60,
         connection_retry_interval: float = 30,
@@ -2402,7 +2406,15 @@ class DataCube(_ProcessGraphAbstraction):
                 method="DataCube.execute_batch()",
             )
 
-        job = cube.create_job(job_options=job_options, validate=validate, auto_add_save_result=False)
+        job = cube.create_job(
+            title=title,
+            description=description,
+            plan=plan,
+            budget=budget,
+            job_options=job_options,
+            validate=validate,
+            auto_add_save_result=False,
+        )
         return job.run_synchronous(
             outputfile=outputfile,
             print=print, max_poll_interval=max_poll_interval, connection_retry_interval=connection_retry_interval